added: Trainee and Admin authentication partially implement

+++ src/main/java/com/cts/dto/LoginResponse.java
@@ -6,5 +6,5 @@ public class LoginResponse {
     private String token;
     private String email;
     private String role;
-    private Long candidateId;
+    private Integer candidateId;
 }
